fix: ui block bug
