+++ packages/umi-build-dev/src/plugins/commands/block/util.js
@@ -195,6 +195,34 @@ export async function gitUpdate(ctx, spinner) {
 }
 
 /**
+ * 打平 children
+ * {
+ *    path:"/user",
+ *    children:[{ path: "/user/list" }]
+ *  }
+ *  --->
+ *  /user /user/list
+ * @param treeData
+ */
+const reduceData = treeData =>
+  treeData.reduce((pre, current) => {
+    const router = pre[current.key];
+    let childrenKeys = {};
+    if (current && current.children) {
+      childrenKeys = reduceData(current.children);
+    }
+
+    if (!router) {
+      pre[current.key] = current;
+    }
+
+    return {
+      ...pre,
+      ...childrenKeys,
+    };
+  }, {});
+
+/**
  * 克隆区块的地址
  * @param {*} ctx
  * @param {*} spinner
@@ -217,3 +245,65 @@ export async function gitClone(ctx, spinner) {
   }
   spinner.succeed();
 }
+
+export const genRouterToTreeData = routes =>
+  routes
+    .map(item =>
+      item.path || item.routes
+        ? {
+            title: item.path,
+            value: item.path,
+            key: item.path,
+            children: genRouterToTreeData(item.routes || []),
+          }
+        : undefined,
+    )
+    .filter(item => item);
+
+/**
+ * 判断路由是否存在
+ * @param {*} path string
+ * @param {*} routes
+ */
+export function routeExists(path, routes = []) {
+  const routerConfig = reduceData(genRouterToTreeData(routes));
+  if (routerConfig[path]) {
+    return true;
+  }
+  return false;
+}
+
+/**
+ *  转化一下
+ *  /user /user/list /user/list/item
+ *  ---->
+ *  {
+ *    path:"/user",
+ *    children:[{ path: "/user/list" }]
+ *  }
+ * @param routes
+ */
+export const depthRouterConfig = routes => {
+  const routerConfig = reduceData(genRouterToTreeData(routes));
+  /**
+   * 这里可以拼接可以减少一次循环
+   */
+  return (
+    Object.keys(routerConfig)
+      .sort((a, b) => a.split('/').length - b.split('/').length + a.length - b.length)
+      .map(key => {
+        key
+          .split('/')
+          .filter(routerKey => routerKey)
+          .forEach((_, index, array) => {
+            const routerKey = array.slice(0, index + 1).join('/');
+            if (routerKey.includes('/')) {
+              delete routerConfig[`/${routerKey}`];
+            }
+          });
+        return routerConfig[key];
+      })
+      // 删除没有 children 的数据
+      .filter(item => item)
+  );
+};
